modified OrderService.java, OrderServiceImpl.java

+++ Backend/Services/Order-service-citymall/src/main/java/com/citymall/orderService/service/OrderService.java
@@ -8,6 +8,6 @@ import java.util.List;
 public interface OrderService {
     OrderResponse createOrder(OrderRequest request);
     List<OrderResponse> findAllOrders();
-    OrderResponse findById(Integer id);
+    OrderResponse findById(Long id);
     List<OrderResponse> findByCustomerId(String customerId);
 }
